Record Access mode mobile support migration
