+++ src/ADempiere/modules/window/WindowType/VuexType.ts
@@ -67,6 +67,7 @@ export interface ContextMenuState {
     contextMenu: IContextMenuData[]
     listDocumentStatus: IListDocumentStatus
     listDocumentAction: IListDocumentAction
+    recordAccess: boolean
 }
 
 // Chat Entries
