+++ src/ADempiere/modules/window/components/ContextMenu/ContextMenuDesktop/template.vue
@@ -145,7 +145,7 @@
     border-radius: 4px;
     -webkit-box-shadow: 0 2px 12px 0 rgba(0, 0, 0, 0.1);
     box-shadow: 0 2px 12px 0 rgba(0, 0, 0, 0.1);
-    max-height: 250px;
+    max-height: 300px;
     max-width: 220px;
     overflow: auto;
   }
