+++ src/ADempiere/modules/window/store/ContextMenu/state.ts
@@ -16,5 +16,6 @@ export const state : ContextMenuState = {
     documentActionsList: [],
     recordId: undefined,
     recordUuid: undefined
-  }
+  },
+  recordAccess: false
 }
