+++ src/ADempiere/shared/lang/es.ts
@@ -309,7 +309,7 @@ export default {
     lockRecord: 'Bloquear Registro',
     noDescription: 'Sin Descripción',
     recordAccess: {
-      actions: 'Acceso a registros',
+      actions: 'Acceso a Registros',
       hideRecord: 'Ocultar Registro',
       recordDisplay: 'Permitir Visualización del Registro',
       isReadonly: 'Solo  Lectura',
